split bulk delete spec into two files to reduce test time
